# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: eSPX Management JSON API
  version: 1.0.0
  description: |
    Machine-readable contract for `/api/v1` JSON routes served by `cmd/management`
    (`internal/management` + `internal/adminapi`). HTMX `/admin/*` HTML routes are excluded.
    HTTP 501 stub routes (deferred dashboards/reports) are excluded — see `internal/adminapi/stub_routes.go` and `internal/openapi/stubs.go`.
servers:
  - url: http://127.0.0.1:8188
    description: Local management (default SERVER_PORT)
components:
  securitySchemes:
    AdminAPIKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Shared admin API key (`ADMIN_API_KEY`).
    SessionCookie:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT session cookie set by `POST /api/v1/auth/login`.
    ConsentHMAC:
      type: apiKey
      in: header
      name: X-Consent-Signature
      description: HMAC-SHA256 of raw request body (`CONSENT_HMAC_SECRET`).
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    GenericObject:
      type: object
      additionalProperties: true
...
    ReportTable:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
...
paths:
...
  /api/v1/postbacks/config:
    get:
      tags: [Integrations]
      summary: GET /api/v1/postbacks/config
      operationId: get_postbacks_config
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/postbacks/config/{campaign_id}:
    put:
      tags: [Integrations]
      summary: PUT /api/v1/postbacks/config/{campaign_id}
      operationId: put_postbacks_config_campaign_id
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/postbacks/dlq:
    get:
      tags: [Integrations]
      summary: GET /api/v1/postbacks/dlq
      operationId: get_postbacks_dlq
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/postbacks/dlq/{id}/retry:
    post:
      tags: [Integrations]
      summary: POST /api/v1/postbacks/dlq/{id}/retry
      operationId: post_postbacks_dlq_id_retry
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/reports/keywords:
    get:
      tags: [Reporting]
      summary: GET /api/v1/reports/keywords
      operationId: get_reports_keywords
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTable'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/reports/placements:
    get:
      tags: [Reporting]
      summary: GET /api/v1/reports/placements
      operationId: get_reports_placements
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTable'
        '401':
          $ref: '#/components/responses/Unauthorized'
...